Initial commit
This is an example that illustrates a bug in Qt 5.15 (Francis Brissette
thinks it was a regression some time between Qt 5.12.5 and Qt 5.15.2)

The bug is in the use of QTextEdit, that exhibits itself when inserting
a large amount of text (4000 lines in this test) into a blank QTextEdit.

When doing this, the application hangs for many seconds (12.7 seconds on
the test machine).